refactor(doc): relocate `git_history` to `rari-doc`
Move the `GIT_HISTORY` static and `git_history()` accessor from
`rari-types::globals` into a new `rari-doc::git_history` module.
`HistoryEntry` stays in `rari-types`. Both existing call sites are
already in `rari-doc`, so this is a pure relocation with no behavior
change.

Sets up a follow-up that needs to compose `git_history` with
`rari-doc::redirects` at load time.
