+++ crates/rari-types/src/globals.rs
@@ -8,7 +8,7 @@ use serde::Deserialize;
 use crate::error::EnvError;
 use crate::locale::Locale;
 use crate::settings::{Deps, Settings};
-use crate::{HistoryEntry, Popularities, globals};
+use crate::{Popularities, globals};
 
 #[inline(always)]
 pub fn content_root() -> &'static Path {
@@ -140,27 +140,6 @@ pub fn json_svg_data_lookup() -> &'static JsonSVGDataLookup {
             serde_json::from_str(&json_str).expect("unable to parse SVGData.json");
         data.elements
     })
-}
-
-pub static GIT_HISTORY: LazyLock<HashMap<PathBuf, HistoryEntry>> = LazyLock::new(|| {
-    let f = content_root().join("_git_history.json");
-    let mut map = if let Ok(json_str) = fs::read_to_string(f) {
-        serde_json::from_str(&json_str).expect("unable to parse l10n json")
-    } else {
-        HashMap::new()
-    };
-    if let Some(translated_root) = content_translated_root() {
-        let f = translated_root.join("_git_history.json");
-        if let Ok(json_str) = fs::read_to_string(f) {
-            let translated: HashMap<PathBuf, HistoryEntry> =
-                serde_json::from_str(&json_str).expect("unable to parse l10n json");
-            map.extend(translated);
-        };
-    }
-    map
-});
-pub fn git_history() -> &'static HashMap<PathBuf, HistoryEntry> {
-    &GIT_HISTORY
 }
 
 pub static POPULARITIES: LazyLock<Popularities> = LazyLock::new(|| {
